Reduce size of main.cpp
